Transplant a stub for all transplant methods
Topology attributes can transplant methods in other classes (mostly the
groups). This transplant, however only happens when a given attribute is
attached to a topology. This causes issues including missing
documentation for the transplanted methods and inonsitent error
messages.

This commit transplants a dummy method fo all transplants. This happens
independently from the unverse creation, so the stubs can appear in the
documentation. The stubs carry the docstring of their counterpart
methods, and raise a NoDataError when called.